Extract tag columns builders to separate files
