+++ middle/middle-pgsql.cpp
@@ -33,6 +33,8 @@ using namespace std;
 #include <libpq-fe.h>
 
 #include "middle-pgsql.hpp"
+#include "jsonb_tags_storage_t.hpp"
+#include "hstore_tags_storage_t.hpp"
 #include "node-persistent-cache.hpp"
 #include "node-ram-cache.hpp"
 #include "options.hpp"
@@ -88,233 +90,6 @@ std::string escape_string(std::string const &in, bool escape)
     }
     return result;
 }
-
-
-class tags_storage_t {
-public:
-    virtual std::string get_column_name()=0;
-
-virtual void pgsql_parse_tags(const char *string, osmium::builder::TagListBuilder & builder) = 0;
-
-virtual std::string encode_tags(osmium::OSMObject const &obj, bool attrs,
-                                       bool escape) = 0;
-
-virtual ~tags_storage_t(){};
-
-};
-
-class jsonb_tags_storage_t : public tags_storage_t{
-
-
-// Decodes a portion of an jsonb literal from postgres */
-// Argument should point to beginning of literal, on return points to delimiter */
-inline const char *decode_upto(const char *src, char *dst)
-{
-    while (*src == ' ')
-        src++;
-    int quoted = (*src == '"');
-    if (quoted)
-        src++;
-
-    while (quoted ? (*src != '"')
-                  : (*src != ',' && *src != '}' && *src != ':')) {
-        if (*src == '\\') {
-            switch (src[1]) {
-            case 'n':
-                *dst++ = '\n';
-                break;
-            case 't':
-                *dst++ = '\t';
-                break;
-            default:
-                *dst++ = src[1];
-                break;
-            }
-            src += 2;
-        } else
-            *dst++ = *src++;
-    }
-    if (quoted)
-        src++;
-    *dst = 0;
-    return src;
-}
-
-public:
-    std::string get_column_name() {return "jsonb";}
-
-void pgsql_parse_tags(const char *string, osmium::builder::TagListBuilder & builder){
-    if (*string++ != '{')
-        return;
-
-    char key[1024];
-    char val[1024];
-
-    while (*string != '}') {
-        string = decode_upto(string, key);
-        // String points to the comma */
-        string++;
-        string = decode_upto(string, val);
-        builder.add_tag(key, val);
-        // String points to the comma or closing '}' */
-        if (*string == ',') {
-            string++;
-        }
-    }
-}
-
-// escape means we return '\N' for copy mode, otherwise we return just nullptr
-std::string encode_tags(osmium::OSMObject const &obj, bool attrs,
-                                       bool escape)
-{
-    std::string result = "{";
-    for (auto const &it : obj.tags()) {
-        result += (fmt("\"%1%\": \"%2%\",") % escape_string(it.key(), escape) % escape_string(it.value(), escape)).str();
-    }
-    if (attrs) {
-        taglist_t extra;
-        extra.add_attributes(obj);
-        for (auto const &it : extra) {
-            result += (fmt("\"%1%\": \"%2%\",") % it.key % escape_string(it.value, escape)).str();
-        }
-    }
-
-    result[result.size() - 1] = '}';
-    return result;
-}
-
-~jsonb_tags_storage_t(){}
-};
-
-class hstore_tags_storage_t : public tags_storage_t {
-
-// Decodes a portion of a hstore literal from postgres */
-// Argument should point to beginning of literal, on return points to delimiter */
-inline const char *decode_upto(const char *src, char *dst)
-{
-    while (*src == ' ')
-        src++;
-    int quoted = (*src == '"');
-    if (quoted)
-        src++;
-
-    while (quoted ? (*src != '"')
-                  : (*src != ',' && *src != '}' && *src != ':')) {
-        if (*src == '\\') {
-            switch (src[1]) {
-            case 'n':
-                *dst++ = '\n';
-                break;
-            case 't':
-                *dst++ = '\t';
-                break;
-            case 'r':
-                *dst++ = '\r';
-                break;
-            case '"':
-                *dst++ = '"';
-                break;
-            case '\\':
-                *dst++ = '\\';
-                break;
-            default:
-                *dst++ = src[1];
-                break;
-            }
-            src += 2;
-        } else
-            *dst++ = *src++;
-    }
-    if (quoted)
-        src++;
-    *dst = 0;
-    return src;
-}
-
-// TODO! copypasted from table.cpp. Extract to orginal lib.
-//create an escaped version of the string for hstore table insert
-void escape4hstore(const char *src, std::string& dst)
-{
-    dst.push_back('"');
-    for (size_t i = 0; i < strlen(src); ++i) {
-        switch (src[i]) {
-            case '\\':
-                dst.append("\\\\\\\\");
-                break;
-            case '"':
-                dst.append("\\\\\"");
-                break;
-            case '\t':
-                dst.append("\\\t");
-                break;
-            case '\r':
-                dst.append("\\\r");
-                break;
-            case '\n':
-                dst.append("\\\n");
-                break;
-            default:
-                dst.push_back(src[i]);
-                break;
-        }
-    }
-    dst.push_back('"');
-}
-
-
-public:
-    std::string get_column_name() {return "hstore";}
-
-void pgsql_parse_tags(const char *string, osmium::builder::TagListBuilder & builder){
-    if (*string != '"')
-        return;
-
-    char key[1024];
-    char val[1024];
-
-    while (strlen(string)) {
-        string = decode_upto(string, key);
-        // Find start of the next string
-        while (*++string!='"') {}
-        string = decode_upto(string, val);
-        builder.add_tag(key, val);
-        // String points to the comma or end */
-        if (*string == ',') {
-            string++;
-        }
-    }
-}
-
-// escape means we return '\N' for copy mode, otherwise we return just nullptr
-std::string encode_tags(osmium::OSMObject const &obj, bool attrs,
-                                       bool escape)
-{
-    std::string result;// = "'";
-    for (auto const &it : obj.tags()) {
-        //result += (fmt("%1%=>%2%,") % escape_string(it.key(), escape) % escape_string(it.value(), escape)).str();
-        escape4hstore(it.key(), result);
-        result += "=>";
-        escape4hstore(it.value(), result);
-        result += ',';
-    }
-    if (attrs) {
-        taglist_t extra;
-        extra.add_attributes(obj);
-        for (auto const &it : extra) {
-            //result += (fmt("%1%=>%2%,") % it.key % escape_string(it.value, escape)).str();
-            escape4hstore(it.key.c_str(), result);
-            result += "=>";
-            escape4hstore(it.value.c_str(), result);
-            result += ',';
-        }
-    }
-
-    result[result.size() - 1] = ' ';
-    return result;
-}
-
-~hstore_tags_storage_t(){}
-};
 
 middle_pgsql_t::table_desc::table_desc()
 : copyMode(0), transactionMode(0), sql_conn(nullptr)
