+++ tests/middle-tests.hpp
@@ -1,7 +1,7 @@
 #ifndef TESTS_MIDDLE_TEST_HPP
 #define TESTS_MIDDLE_TEST_HPP
 
-#include "middle.hpp"
+#include "middle/middle.hpp"
 
 // tests that a single node can be set and retrieved. returns 0 on success.
 int test_node_set(middle_t *mid);
