+++ tests/test-hstore-match-only.cpp
@@ -16,11 +16,11 @@ The tags of inteest are specified in hstore-match-only.style
 
 #include "osmtypes.hpp"
 #include "osmdata.hpp"
-#include "middle.hpp"
+#include "middle/middle.hpp"
 #include "output-pgsql.hpp"
 #include "options.hpp"
-#include "middle-pgsql.hpp"
-#include "middle-ram.hpp"
+#include "middle/middle-pgsql.hpp"
+#include "middle/middle-ram.hpp"
 #include "taginfo_impl.hpp"
 
 #include <sys/types.h>
