+++ tests/test-output-multi-point-multi-table.cpp
@@ -11,7 +11,7 @@
 #include "osmdata.hpp"
 #include "output-multi.hpp"
 #include "options.hpp"
-#include "middle-pgsql.hpp"
+#include "middle/middle-pgsql.hpp"
 #include "taginfo_impl.hpp"
 
 #include <sys/types.h>
